requirements.txt
# 9LMNTS STUDIO - Automation System Requirements
# Core dependencies for secure automation system

# Configuration and Environment
python-dotenv==1.0.0
requests==2.31.0
openai==1.6.1

# Database and Storage
sqlite3

# Error Handling and Logging
typing-extensions==4.8.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Advanced Features
# Uncomment if needed for enhanced functionality
# aiosmtplib>=1.0.0
# aiohttp>=3.8.0
# async-timeout>=4.0.0

# Security and Validation
# cryptography>=41.0.0
# pyopenssl>=23.0.0
